feat: arquitetura feita
