+++ src/services/useCases/checkUnfollowAndFollow/checkUnfollowAndFollowUseCase.ts
@@ -1,0 +1,22 @@
+import { fetchUserFollowData } from "../fetchUserFollowData/fetchUserFollowDataUseCase";
+
+export async function checkUnfollowAndFollow(userName: string): Promise<string[] | null> {
+    const userToFollow: string[] = [];
+    
+    const userFollowData = await fetchUserFollowData(userName);
+    if (!userFollowData) 
+        return null;
+
+    const following = userFollowData.following;
+    const followers = userFollowData.followers;
+
+    followers.forEach(user => {
+    if (!following.includes(user))
+        userToFollow.push(user);
+    });
+    
+    if  ( userToFollow.length === 0) 
+        return ['No user to follow'];
+
+    return userToFollow;
+}
