--- src/utility/qtdePG.js
@@ -1,4 +1,0 @@
-const perPage = 500; // Número de resultados por página
-let page = 1; // Inicializa a página para o primeiro loop
-
-module.exports = { perPage, page };
